// src/lib/pe/types.ts
export interface DialogItem {
  style: number;
  exStyle: number;
  x: number;
  y: number;
  cx: number;
  cy: number;
  id: number;
  className: string;
  text: string;
  titleOrdinal: number | null;
}

export interface DialogFont {
  pointSize: number;
  weight?: number;
  italic?: boolean;
  typeface: string;
}

export interface DialogTemplate {
  style: number;
  exStyle: number;
  x: number;
  y: number;
  cx: number;
  cy: number;
  title: string;
  className: string | null;
  menuName: string | null;
  font: DialogFont | null;
  items: DialogItem[];
}

export interface DialogResult {
  id: number | null;
  name: string | null;
  languageId: number;
  dialog: DialogTemplate;
}